Moved regex & helper functions to individual files
This was made in an attempt to make the project easier to navigate
