+++ fortls/constants.py
@@ -1,0 +1,3 @@
+import sys
+
+PY3K = sys.version_info >= (3, 0)
