+++ fortls/langserver.py
@@ -11,13 +11,13 @@ import traceback
 from multiprocessing import Pool
 from pathlib import Path
 
+# Local modules
+from fortls.helper_functions import expand_name
 from fortls.intrinsics import (
     get_intrinsic_keywords,
     load_intrinsics,
     set_lowercase_intrinsics,
 )
-
-# Local modules
 from fortls.jsonrpc import path_from_uri, path_to_uri
 from fortls.objects import (
     CLASS_TYPE_ID,
@@ -31,30 +31,31 @@ from fortls.objects import (
     climb_type_tree,
     find_in_scope,
     find_in_workspace,
+    fortran_ast,
     fortran_var,
+    get_paren_level,
     get_use_tree,
     get_var_stack,
     set_keyword_ordering,
 )
-from fortls.parse_fortran import (
+from fortls.parse_fortran import fortran_file, get_line_context, process_file
+from fortls.regex_patterns import (
     DQ_STRING_REGEX,
     LOGICAL_REGEX,
     NUMBER_REGEX,
     SQ_STRING_REGEX,
-    expand_name,
-    fortran_ast,
-    fortran_file,
-    get_line_context,
-    get_paren_level,
-    process_file,
 )
 
 log = logging.getLogger(__name__)
 # Global regexes
 FORTRAN_EXT_REGEX = re.compile(r"\.F(77|90|95|03|08|OR|PP)?$", re.I)
+# TODO: I think this can be replaced by fortls.regex_patterns
 INT_STMNT_REGEX = re.compile(r"^[ ]*[a-z]*$", re.I)
+# TODO: I think this can be replaced by fortls.regex_patterns type & class
 TYPE_DEF_REGEX = re.compile(r"[ ]*(TYPE|CLASS)[ ]*\([a-z0-9_ ]*$", re.I)
+# TODO: I think this can be replaced by fortls.regex_patterns
 SCOPE_DEF_REGEX = re.compile(r"[ ]*(MODULE|PROGRAM|SUBROUTINE|FUNCTION)[ ]+", re.I)
+# TODO: I think this can be replaced by fortls.regex_patterns END_REGEx
 END_REGEX = re.compile(
     r"[ ]*(END)( |MODULE|PROGRAM|SUBROUTINE|FUNCTION|TYPE|DO|IF|SELECT)?", re.I
 )
